refactor: move ValidationException to utility.exceptions
