+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Internal Changes
 
 Implemented `ClientSessionManager` for client-side session management.
+Implemented credential verification in `BuyStockController`.
+Moved `ValidationException` into a separate class in utility.
 
 ### Bug Fixes
