+++ src/main/java/use_case/execute_buy/ExecuteBuyDataAccessInterface.java
@@ -1,7 +1,7 @@
 package use_case.execute_buy;
 
 import entity.User;
-import utility.ValidationException;
+import utility.exceptions.ValidationException;
 
 public interface ExecuteBuyDataAccessInterface {
     User getUserWithCredential(String credential) throws ValidationException;
