+++ src/main/java/utility/exceptions/ValidationException.java
@@ -1,4 +1,4 @@
-package utility;
+package utility.exceptions;
 
 // custom exception for validation error
 public class ValidationException extends Exception {
