+++ src/test/java/use_case/execute_buy/ExecuteBuyInteractorTest.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import utility.ValidationException;
+import utility.exceptions.ValidationException;
 
 class ExecuteBuyInteractorTest {
